Build 003
Totally revamp and change the game. The previous structure left too
little room for freedom and how the game can expand. The main loop was
too boring, with too many clicks, and not enough growth.

The farm has been expanded to be a set of tiles, of which allow you to
zoom in on them and perform actions. The farm will need to be cleared,
cleaned, and managed.

This gives way to a lot more resources you can get on the farm to start.

Add in activities, which allow the player to do mulitple things at the
same time, and these actions take real time to complete.

Transitioning a lot of the code to use classes. This will break saving
and loading until I fix loading. I'm not the biggest fan of classes, but
it simplifies a lot of the structure.